+++ ICT397_Assignment/ICT397_Assignment/Control.cpp
@@ -66,8 +66,8 @@ void Control::OpenGLKeyboardUpFunc(unsigned char key, int x, int y){
 void Control::OpenGLMouseMovementFunc(int x, int y){
 	if (x != mouse_x_org || y != mouse_y_org){
 		cam->SetCameraRotation(
-			((float)x - ((float)screen_width/2)),
-			((float)y - ((float)screen_height/2))
+			((double)x - ((double)screen_width/2)),
+			((double)y - ((double)screen_height/2))
 			);
 		cam->SetIsRotating(true);
 		glutWarpPointer(screen_width / 2, screen_height / 2);
